requirements.txt
flask
pandas
numpy
python-dotenv
gunicorn
fastapi
uvicorn
python-multipart
mailjet_rest